Node: Reuse servers between tests.
This redues the test runtime from ~29seconds to ~6seconds, and reduces
the load on the testing machine.
