+++ node/tests/SharedTests.ts
@@ -1,0 +1,222 @@
+import { expect, it } from "@jest/globals";
+import { v4 as uuidv4 } from "uuid";
+import { Client, GetAndSetRandomValue } from "./TestUtilities";
+
+type BaseClient = {
+    set: (
+        key: string,
+        value: string,
+        options?: any
+    ) => Promise<string | "OK" | null>;
+    get: (key: string) => Promise<string | null>;
+    customCommand: (
+        commandName: string,
+        args: string[]
+    ) => Promise<string | string[] | number | null>;
+};
+
+export function runBaseTests<Context>(config: {
+    init: () => Promise<{ context: Context; client: BaseClient }>;
+    close: (context: Context) => void;
+    timeout?: number;
+}) {
+    runCommonTests(config);
+
+    const runTest = async (test: (client: BaseClient) => Promise<void>) => {
+        const { context, client } = await config.init();
+
+        try {
+            await test(client);
+        } finally {
+            config.close(context);
+        }
+    };
+
+    it(
+        "set with return of old value works",
+        async () => {
+            await runTest(async (client: BaseClient) => {
+                const key = uuidv4();
+                // Adding random repetition, to prevent the inputs from always having the same alignment.
+                const value = uuidv4() + "0".repeat(Math.random() * 7);
+
+                let result = await client.set(key, value);
+                expect(result).toEqual("OK");
+
+                result = await client.set(key, "", {
+                    returnOldValue: true,
+                });
+                expect(result).toEqual(value);
+
+                result = await client.get(key);
+                expect(result).toEqual("");
+            });
+        },
+        config.timeout
+    );
+
+    it(
+        "conditional set works",
+        async () => {
+            await runTest(async (client: BaseClient) => {
+                const key = uuidv4();
+                // Adding random repetition, to prevent the inputs from always having the same alignment.
+                const value = uuidv4() + "0".repeat(Math.random() * 7);
+                let result = await client.set(key, value, {
+                    conditionalSet: "onlyIfExists",
+                });
+                expect(result).toEqual(null);
+
+                result = await client.set(key, value, {
+                    conditionalSet: "onlyIfDoesNotExist",
+                });
+                expect(result).toEqual("OK");
+                expect(await client.get(key)).toEqual(value);
+
+                result = await client.set(key, "foobar", {
+                    conditionalSet: "onlyIfDoesNotExist",
+                });
+                expect(result).toEqual(null);
+
+                result = await client.set(key, "foobar", {
+                    conditionalSet: "onlyIfExists",
+                });
+                expect(result).toEqual("OK");
+
+                expect(await client.get(key)).toEqual("foobar");
+            });
+        },
+        config.timeout
+    );
+
+    it(
+        "custom command works",
+        async () => {
+            await runTest(async (client: BaseClient) => {
+                const key = uuidv4();
+                // Adding random repetition, to prevent the inputs from always having the same alignment.
+                const value = uuidv4() + "0".repeat(Math.random() * 7);
+                const setResult = await client.customCommand("SET", [
+                    key,
+                    value,
+                ]);
+                expect(setResult).toEqual("OK");
+                const result = await client.customCommand("GET", [key]);
+                expect(result).toEqual(value);
+            });
+        },
+        config.timeout
+    );
+}
+
+export function runCommonTests<Context>(config: {
+    init: () => Promise<{ context: Context; client: Client }>;
+    close: (context: Context) => void;
+    timeout?: number;
+}) {
+    const runTest = async (test: (client: Client) => Promise<void>) => {
+        const { context, client } = await config.init();
+
+        try {
+            await test(client);
+        } finally {
+            config.close(context);
+        }
+    };
+
+    it(
+        "set and get flow works",
+        async () => {
+            await runTest((client: Client) => GetAndSetRandomValue(client));
+        },
+        config.timeout
+    );
+
+    it(
+        "can handle non-ASCII unicode",
+        async () => {
+            await runTest(async (client: Client) => {
+                const key = uuidv4();
+                const value = "שלום hello 汉字";
+                await client.set(key, value);
+                const result = await client.get(key);
+                expect(result).toEqual(value);
+            });
+        },
+        config.timeout
+    );
+
+    it(
+        "get for missing key returns null",
+        async () => {
+            await runTest(async (client: Client) => {
+                const result = await client.get(uuidv4());
+
+                expect(result).toEqual(null);
+            });
+        },
+        config.timeout
+    );
+
+    it(
+        "get for empty string",
+        async () => {
+            await runTest(async (client: Client) => {
+                const key = uuidv4();
+                await client.set(key, "");
+                const result = await client.get(key);
+
+                expect(result).toEqual("");
+            });
+        },
+        config.timeout
+    );
+
+    it(
+        "send very large values",
+        async () => {
+            await runTest(async (client: Client) => {
+                const WANTED_LENGTH = Math.pow(2, 16);
+                const getLongUUID = () => {
+                    let id = uuidv4();
+                    while (id.length < WANTED_LENGTH) {
+                        id += uuidv4();
+                    }
+                    return id;
+                };
+                const key = getLongUUID();
+                const value = getLongUUID();
+                await client.set(key, value);
+                const result = await client.get(key);
+
+                expect(result).toEqual(value);
+            });
+        },
+        config.timeout
+    );
+
+    it(
+        "can handle concurrent operations",
+        async () => {
+            await runTest(async (client: Client) => {
+                const singleOp = async (index: number) => {
+                    if (index % 2 === 0) {
+                        await GetAndSetRandomValue(client);
+                    } else {
+                        const result = await client.get(uuidv4());
+                        expect(result).toEqual(null);
+                    }
+                };
+
+                const operations: Promise<void>[] = [];
+
+                for (let i = 0; i < 100; ++i) {
+                    operations.push(singleOp(i));
+                }
+
+                await Promise.all(operations);
+            });
+        },
+        config.timeout
+    );
+}
